+++ packages/default-client/CHANGELOG.md
@@ -1,5 +1,12 @@
 # fuzd-default-client
 
+## 0.1.13
+
+### Patch Changes
+
+- Updated dependencies
+  - fuzd-scheduler@0.2.0
+
 ## 0.1.12
 
 ### Patch Changes