+++ packages/tlock-decrypter/CHANGELOG.md
@@ -1,5 +1,16 @@
 # fuzd-tlock-decrypter
 
+## 0.2.0
+
+### Minor Changes
+
+- batch of tx in scheduler
+
+### Patch Changes
+
+- Updated dependencies
+  - fuzd-scheduler@0.2.0
+
 ## 0.1.9
 
 ### Patch Changes